refactor: 미사용 메소드 제거
