+++ src/test/java/org/devcourse/resumeme/business/event/controller/EventControllerTest.java
@@ -243,71 +243,6 @@ class EventControllerTest extends ControllerUnitTest {
     }
 
     @Test
-    void 첨삭_이벤트_신청을_반려한다() throws Exception {
-        // given
-        EventRejectRequest request = new EventRejectRequest("이력서 작성 양이 너무 적습니다");
-        doNothing().when(eventService).reject(new EventReject(1L, 1L, "이력서 작성 양이 너무 적습니다"));
-
-        // when
-        ResultActions result = mvc.perform(patch("/api/v1/events/{eventId}/mentee/{menteeId}", 1L, 1L)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJson(request)));
-
-        // then
-        result
-                .andExpect(status().isOk())
-                .andDo(
-                        document("event/reject",
-                                getDocumentRequest(),
-                                pathParameters(
-                                        parameterWithName("eventId").description("이벤트 아이디"),
-                                        parameterWithName("menteeId").description("반려시키고 싶은 멘티 아이디")
-                                ),
-                                exceptionResponse(
-                                        List.of(EVENT_REJECTED.name())
-                                ),
-                                requestFields(
-                                        fieldWithPath("rejectMessage").type(STRING).description("이벤트 신청 반려 사유").optional()
-                                )
-                        )
-                );
-    }
-
-    @Test
-    @WithMockCustomUser
-    void 첨삭_이벤트_리뷰를_완료한다() throws Exception {
-        // given
-        Long eventId = 1L;
-        Long resumeId = 1L;
-        CompleteEventRequest request = new CompleteEventRequest("좋은 이력서에요.");
-
-        // when
-        ResultActions result = mvc.perform(patch("/api/v1/events/{eventId}/resumes/{resumeId}/complete", eventId, resumeId)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJson(request)));
-
-        // then
-        result
-                .andExpect(status().isOk())
-                .andDo(
-                        document("event/completeReview",
-                                getDocumentRequest(),
-                                getDocumentResponse(),
-                                pathParameters(
-                                        parameterWithName("eventId").description("이벤트 아이디"),
-                                        parameterWithName("resumeId").description("이력서 아이디")
-                                ),
-                                exceptionResponse(
-                                        List.of(EVENT_NOT_FOUND.name(), RESUME_NOT_FOUND.name())
-                                ),
-                                requestFields(
-                                        fieldWithPath("comment").type(STRING).description("이벤트 리뷰 내용")
-                                )
-                        )
-                );
-    }
-
-    @Test
     @WithMockCustomUser(role = "ROLE_MENTOR")
     void 첨삭_이벤트_상세_조회할수있다() throws Exception {
         // given
