+++ src/test/java/org/devcourse/resumeme/business/event/service/EventServiceTest.java
@@ -4,13 +4,11 @@ import org.devcourse.resumeme.business.event.domain.Event;
 import org.devcourse.resumeme.business.event.domain.EventInfo;
 import org.devcourse.resumeme.business.event.domain.EventTimeInfo;
 import org.devcourse.resumeme.business.event.exception.EventException;
-import org.devcourse.resumeme.business.user.domain.mentee.RequiredInfo;
-import org.devcourse.resumeme.business.user.domain.mentor.Mentor;
+import org.devcourse.resumeme.business.event.repository.EventRepository;
 import org.devcourse.resumeme.business.user.domain.Provider;
 import org.devcourse.resumeme.business.user.domain.Role;
-import org.devcourse.resumeme.business.event.repository.EventRepository;
-import org.devcourse.resumeme.business.event.service.vo.AcceptMenteeToEvent;
-import org.devcourse.resumeme.business.event.service.vo.EventReject;
+import org.devcourse.resumeme.business.user.domain.mentee.RequiredInfo;
+import org.devcourse.resumeme.business.user.domain.mentor.Mentor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
@@ -22,28 +20,20 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
-import java.util.Random;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Consumer;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class EventServiceTest {
-
-    @Mock
-    private MenteeToEventService menteeToEventService;
 
     @Mock
     private EventRepository eventRepository;
@@ -117,78 +107,6 @@ class EventServiceTest {
         // when & then
         assertThatThrownBy(() -> eventService.create(event))
                 .isInstanceOf(EventException.class);
-    }
-
-
-    @Test
-    void 여러스레드를_이용하여_선착순으로_진행한_이벤트참여신청에_최대_참여수만큼_성공한다() throws InterruptedException {
-        // given
-        EventInfo openEvent = EventInfo.open(3, "제목", "내용");
-        EventTimeInfo eventTimeInfo = EventTimeInfo.onStart(LocalDateTime.now(), LocalDateTime.now().plusHours(1L), LocalDateTime.now().plusHours(2L));
-        Event event = new Event(openEvent, eventTimeInfo, mentor, List.of());
-
-        given(eventRepository.findWithLockById(1L)).willReturn(Optional.of(event));
-        given(menteeToEventService.getApplyEventCount(1L)).willReturn(0L);
-
-        // when
-        for (int i = 0; i < executeCount; i++) {
-            executeThread((number) -> eventService.acceptMentee(new AcceptMenteeToEvent(1L, Long.valueOf(number), 1L)), i);
-        }
-
-        countDownLatch.await();
-
-        // then
-        assertThat(successCount.get()).isEqualTo(3);
-        assertThat(failCount.get()).isEqualTo(executeCount - 3);
-    }
-
-    private void executeThread(Consumer<Integer> service, int number) {
-            executorService.execute(() -> {
-                try {
-                    Thread.sleep(new Random().nextLong(100));
-
-                    service.accept(number);
-
-                    successCount.getAndIncrement();
-                } catch (Exception e) {
-                    failCount.getAndIncrement();
-                    System.out.println(e.getMessage());
-                }
-                countDownLatch.countDown();
-            });
-    }
-
-    @Test
-    void 이벤트_신청을_반려에_성공한다() {
-        // given
-        EventInfo openEvent = EventInfo.open(3, "제목", "내용");
-        EventTimeInfo eventTimeInfo = EventTimeInfo.onStart(LocalDateTime.now(), LocalDateTime.now().plusHours(1L), LocalDateTime.now().plusHours(2L));
-        Event event = new Event(openEvent, eventTimeInfo, mentor, List.of());
-        event.acceptMentee(1L, 1L);
-
-        EventReject reject = new EventReject(1L, 1L, "이력서 양이 너무 적습니다");
-        lenient().when(eventRepository.findWithApplicantsById(1L)).thenReturn(Optional.of(event));
-
-        // when
-        eventService.reject(reject);
-
-        // then
-        verify(eventRepository).findWithApplicantsById(1L);
-    }
-
-    @Test
-    void 이벤트_신청을_반려에_실패한다_신청한_멘티가_없음() {
-        // given
-        EventInfo openEvent = EventInfo.open(3, "제목", "내용");
-        EventTimeInfo eventTimeInfo = EventTimeInfo.onStart(LocalDateTime.now(), LocalDateTime.now().plusHours(1L), LocalDateTime.now().plusHours(2L));
-        Event event = new Event(openEvent, eventTimeInfo, mentor, List.of());
-        event.acceptMentee(1L, 1L);
-
-        EventReject reject = new EventReject(1L, 2L, "이력서 양이 너무 적습니다");
-        given(eventRepository.findWithApplicantsById(1L)).willReturn(Optional.of(event));
-
-        // when & then
-        assertThatThrownBy(() -> eventService.reject(reject)).isInstanceOf(EventException.class);
     }
 
 }
