+++ src/test/java/org/devcourse/resumeme/business/event/service/MenteeToEventServiceTest.java
@@ -4,7 +4,9 @@ import org.devcourse.resumeme.business.event.domain.Event;
 import org.devcourse.resumeme.business.event.domain.EventInfo;
 import org.devcourse.resumeme.business.event.domain.EventTimeInfo;
 import org.devcourse.resumeme.business.event.domain.MenteeToEvent;
+import org.devcourse.resumeme.business.event.repository.EventRepository;
 import org.devcourse.resumeme.business.event.repository.MenteeToEventRepository;
+import org.devcourse.resumeme.business.event.service.vo.AcceptMenteeToEvent;
 import org.devcourse.resumeme.business.user.domain.Provider;
 import org.devcourse.resumeme.business.user.domain.Role;
 import org.devcourse.resumeme.business.user.domain.mentee.Mentee;
@@ -21,7 +23,14 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
+import java.util.Random;
 import java.util.Set;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Consumer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
@@ -35,20 +44,36 @@ class MenteeToEventServiceTest {
     private MenteeToEventService menteeToEventService;
 
     @Mock
+    private EventRepository eventRepository;
+
+    @Mock
     private MenteeToEventRepository menteeToEventRepository;
 
     private Mentor mentorOne;
 
     private Mentor mentorTwo;
 
-    private Mentee mentee;
-
     private Event eventOne;
 
     private Event eventTwo;
 
+    private ExecutorService executorService;
+
+    private CountDownLatch countDownLatch;
+
+    private AtomicInteger successCount;
+
+    private AtomicInteger failCount;
+
+    int executeCount;
+
     @BeforeEach
     void setUp() {
+        executeCount = 10;
+        executorService = Executors.newFixedThreadPool(3);
+        countDownLatch = new CountDownLatch(executeCount);
+        successCount = new AtomicInteger();
+        failCount = new AtomicInteger();
 
         mentorOne =  Mentor.builder()
                 .id(1L)
@@ -74,18 +99,6 @@ class MenteeToEventServiceTest {
                 .careerYear(5)
                 .build();
 
-        mentee = Mentee.builder()
-                .id(1L)
-                .imageUrl("menteeimage.png")
-                .provider(Provider.valueOf("KAKAO"))
-                .email("backdong1@kakao.com")
-                .refreshToken("ddefweferfrte")
-                .requiredInfo(new RequiredInfo("김백둥", "백둥둥", "01022223722", Role.ROLE_MENTEE))
-                .interestedPositions(Set.of())
-                .interestedFields(Set.of())
-                .introduce(null)
-                .build();
-
         EventInfo openEvent = EventInfo.book(3, "제목", "내용");
         EventTimeInfo eventTimeInfo = EventTimeInfo.book(LocalDateTime.now(), LocalDateTime.now().plusHours(1L), LocalDateTime.now().plusHours(2L), LocalDateTime.now().plusHours(4L));
         eventOne = new Event(openEvent, eventTimeInfo, mentorOne, List.of());
@@ -105,6 +118,44 @@ class MenteeToEventServiceTest {
 
         //then
         assertThat(eventsRelatedToResume.size()).isEqualTo(2);
+    }
+
+    @Test
+    void 여러스레드를_이용하여_선착순으로_진행한_이벤트참여신청에_최대_참여수만큼_성공한다() throws InterruptedException {
+        // given
+        EventInfo openEvent = EventInfo.open(3, "제목", "내용");
+        EventTimeInfo eventTimeInfo = EventTimeInfo.onStart(LocalDateTime.now(), LocalDateTime.now().plusHours(1L), LocalDateTime.now().plusHours(2L));
+        Event event = new Event(openEvent, eventTimeInfo, mentorOne, List.of());
+
+        given(eventRepository.findWithLockById(1L)).willReturn(Optional.of(event));
+        given(menteeToEventRepository.findByMenteeId(1L)).willReturn(List.of(new MenteeToEvent(event, 1L, 1L)));
+
+        // when
+        for (int i = 0; i < executeCount; i++) {
+            executeThread((number) -> menteeToEventService.acceptMentee(new AcceptMenteeToEvent(1L, Long.valueOf(number), 1L)), i);
+        }
+
+        countDownLatch.await();
+
+        // then
+        assertThat(successCount.get()).isEqualTo(3);
+        assertThat(failCount.get()).isEqualTo(executeCount - 3);
+    }
+
+    private void executeThread(Consumer<Integer> service, int number) {
+        executorService.execute(() -> {
+            try {
+                Thread.sleep(new Random().nextLong(100));
+
+                service.accept(number);
+
+                successCount.getAndIncrement();
+            } catch (Exception e) {
+                failCount.getAndIncrement();
+                System.out.println(e.getMessage());
+            }
+            countDownLatch.countDown();
+        });
     }
 
 }
